Schedule internals (#25)
* Exposing internals of isl_schedule and isl_schedule_[tree|band|node] for moving to C++

* Expose more functions
